python/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.108.0
frozenlist==1.4.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-tools==1.60.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
hyperframe==6.0.1
icecream==2.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.327
langsmith==0.0.56
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.6.1
orjson==3.9.10
packaging==23.2
portalocker==2.8.2
protobuf==4.25.1
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.16.1
python-dotenv==1.0.0
python-multipart==0.0.6
pywin32==306; platform_system=="Windows"
PyYAML==6.0.1
qdrant-client==1.7.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.32.0.post1
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
